refactor(client): consolidate dead feature dirs into unified feature dirs
Move component files from standalone directories into their unified
feature directories to match the routing consolidation already in place:

- agent-comms/{agent-comms,agent-network-vis,agent-network-3d}.component.ts
  → comms/
- feed/live-feed.component.ts → comms/
- brain-viewer/{brain-viewer.component,brain-viewer.component.spec}.ts
  → memory/
- memory-browser/memory-browser.component.ts → memory/
- contacts/contact-list.component.ts → settings-integrations/

Update import paths in unified-comms, unified-memory, and
settings-integrations components from cross-directory to local (./).

Old directories are now empty and deleted. Routes were already redirected
in a prior change — no app.routes.ts changes needed.

Grep proof (non-redirect imports resolved):
- agent-comms: only unified-comms.component.ts (now ./agent-comms.component)
- feed: only unified-comms.component.ts (now ./live-feed.component)
- brain-viewer: only unified-memory.component.ts (now ./brain-viewer.component)
- memory-browser: only unified-memory.component.ts (now ./memory-browser.component)
- contacts: only settings-integrations.component.ts (now ./contact-list.component)
