+++ client/src/app/features/comms/agent-network-3d.component.ts
