+++ client/src/app/features/comms/agent-network-vis.component.ts
