+++ client/src/app/features/memory/brain-viewer.component.ts
