+++ client/src/app/features/memory/memory-browser.component.ts
